modified login-sevice files
