+++ angular-frontend/src/app/services/user.service.ts
@@ -7,7 +7,7 @@ import { Observable } from 'rxjs';
 })
 export class UserService {
 
-  private userUrl = 'http://localhost:8080/api/test/user';
+  private userUrl = 'http://localhost:8083/api/test/user';
   // private pmUrl = 'http://localhost:8080/api/test/pm';
   // private adminUrl = 'http://localhost:8080/api/test/admin';
