+++ login-service/src/main/java/com/stackroute/LoginServiceApplication.java
@@ -1,9 +1,11 @@
-package com.stackroute.knowledgehub.loginservice;
+package com.stackroute;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
 
 @SpringBootApplication
+@EnableEurekaClient
 public class LoginServiceApplication  {
 
 	public static void main(String[] args) {
