+++ login-service/src/main/java/com/stackroute/controller/TestRestAPI.java
@@ -1,5 +1,6 @@
-package com.stackroute.knowledgehub.loginservice.controller;
+package com.stackroute.controller;
 
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -8,6 +9,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class TestRestAPI {
     @GetMapping("/api/test/user")
+    @PreAuthorize("authenticated")
     public String userAccess() {
         return "Authenticated Contents!";
     }
