--- login-service/src/main/java/com/stackroute/knowledgehub/loginservice/domain/User.java
@@ -1,62 +1,0 @@
-package com.stackroute.knowledgehub.loginservice.domain;
-
-import javax.persistence.*;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.Size;
-@Entity
-@Table(name = "users", uniqueConstraints = {
-        @UniqueConstraint(columnNames = {
-                "username"
-        })
-
-})
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @NotBlank
-    @Size(min=3, max = 50)
-    private String username;
-
-
-    @NotBlank
-    @Size(min=6, max = 100)
-    private String password;
-
-
-    public User() {}
-
-    public User(String name, String username, String email, String password) {
-        //    this.name = name;
-        this.username = username;
-        //    this.email = email;
-        this.password = password;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-}
