+++ login-service/src/main/java/com/stackroute/security/jwt/JwtAuthEntryPoint.java
@@ -1,4 +1,4 @@
-package com.stackroute.knowledgehub.loginservice.security.jwt;
+package com.stackroute.security.jwt;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
