+++ login-service/src/main/java/com/stackroute/security/jwt/JwtProvider.java
@@ -1,6 +1,6 @@
-package com.stackroute.knowledgehub.loginservice.security.jwt;
+package com.stackroute.security.jwt;
 
-import com.stackroute.knowledgehub.loginservice.security.services.UserPrinciple;
+import com.stackroute.security.services.UserPrinciple;
 import io.jsonwebtoken.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -8,7 +8,6 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Component;
 
-import java.security.SignatureException;
 import java.util.Date;
 @Component
 public class JwtProvider {
