--- login-service/src/main/resources/application.properties
@@ -1,16 +1,0 @@
-spring.datasource.url=jdbc:mysql://localhost:3306/LoginDB?useSSL=false&createDatabaseIfNotExist=true&allowPublicKeyRetrieval=true
-#spring.datasource.url=jdbc:mysql://localhost:3306/LoginDB
-spring.datasource.username=root
-spring.datasource.password=root123
-spring.jpa.generate-ddl=true
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.jpa.hibernate.ddl-auto=create
-spring.jpa.hibernate.show-sql= true
-spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5Dialect
-
-
-# App Properties
-knowledgehub.jwtSecret=jwtSecretKey
-knowledgehub.jwtExpiration=86400
-
-server.port=8083
